molesec-profile-data-service/requirements.txt
APScheduler==3.10.0
Django==4.1.5
argon2-cffi==21.3.0
asgiref==3.6.0
astroid==2.13.3
cachetools==5.3.0
click==8.1.3
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
djangorestframework-api-key==2.3.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
eventlet==0.30.2
gunicorn==20.0.4
psycopg2-binary==2.9.5
pylint==2.15.10
python-dotenv==0.21.1
pytz==2022.7.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
whitenoise==6.3.0
wrapt==1.14.1
sentry-sdk==1.19.1
